medicalreportanalyzer/pyproject.toml
[project]
name = "medicalreportanalyzer"
version = "0.1.0"
description = "MedicalReportAnalyzer using crewAI - Analyzes medical report images and provides patient-friendly health summaries"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[google-genai,tools]==1.14.3",
    "google-generativeai>=0.8.0",
    "streamlit>=1.57.0",
]

[project.scripts]
medicalreportanalyzer = "medicalreportanalyzer.main:run"
run_crew = "medicalreportanalyzer.main:run"
train = "medicalreportanalyzer.main:train"
replay = "medicalreportanalyzer.main:replay"
test = "medicalreportanalyzer.main:test"
run_with_trigger = "medicalreportanalyzer.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
